+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2025-07-17
+
+- Added a new `Null_Cursor` class that does not store the cursor in the database.
+- Deprecated the `Cursor` class, which now extends the `Alley\WP_Bulk_Task\Cursor\Option_Cursor` class.
+
 ## 1.0.0 - 2024-08-13
 
 - Stable release.